Search the key in the parsed json
Testing the ogClient I found that if a value of the json match a key the
ogClient has an exception. For example:

	body = "... shell/run {"run": "fdisk -l"} ..."

	CURRENT
	Enters in
	if "disk" in body:...
	if "run" in body:...

	EXPECTED
	Enters in
	if "run" in body:...

This commit changes the behaviour to search for the keys in the
dictionary returned by json.loads() instead of searching in the raw
string. This way the ogClient looks for the keys without searching in
the values.