#1065 Fix windows not reattempting refused connections
Windows does not report a refused connection the same way as Linux.
Unsuccesful connect socket will be kept in the exceptfds, and won't
be in the readable nor writable fds. The socket in this state will
have SO_ERROR set to ECONNREFUSED.

On the other hand, Linux does not use exceptfds for such case.