image_create: add legacy ogCopyEfiBootloader
Use legacy script that saves the Windows-specific content from the ESP
to the image target filesystem.

Current image restore solution from OpenGnsys scripts expect the EFI
partition to be stored in the target system partition. (Only for Windows
10)

For example, storing the ESP in the NTFS partition of a Windows image.

Expect use of bash script ogCopyEfiBootloader until further
integration is merged.