live: clear ogbrowser logs before image_restore
Clears content of blue text areas in the real time log view before
executing a restore image operation.

Adds private function _ogbrowser_clear_logs, this function writes to a
couple of text files present in the ogLive environment.
The contents of this file are printed out to the blue text areas
in the "real time log" view.